requirements.txt
Django==2.0.5
djangorestframework==3.11.0
Pillow==7.1.1
pytz==2019.3
